refactor: remove coap prefix from codec files
